% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleEigengenes.R
\name{AvgModuleExpr}
\alias{AvgModuleExpr}
\title{AverageModuleExpr}
\usage{
AvgModuleExpr(seurat_obj, n_genes = 25, wgcna_name = NULL, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Computes module eigengenes for all WGCNA co-expression modules
}
