% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{BinPseudotime}
\alias{BinPseudotime}
\title{BinPseudotime}
\usage{
BinPseudotime(seurat_obj, n_bins = 50, pseudotime_col = "pseudotime")
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{n_bins}{number of hub genes to use in the UMAP computation}

\item{pseudotime_col}{name of the column in seurat_obj@meta.data containing pseudotime information}
}
\description{
Makes evenly-spaced bins of cells along a pseudotime trajectory.
}
