% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ComputeModuleEigengene}
\alias{ComputeModuleEigengene}
\title{ComputeModuleEigengene}
\usage{
ComputeModuleEigengene(
  seurat_obj,
  cur_mod,
  modules,
  group.by.vars = NULL,
  verbose = TRUE,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{cur_mod}{name of a module found in seurat_obj@misc[\link{seurat_obj@misc$active_wgcna}]$wgcna_net$colors}
}
\description{
Internal helper function that computes module eigengene for a single module.
}
\examples{
ConstructNetwork(pbmc)
}
\keyword{scRNA-seq}
