% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ComputeROC}
\alias{ComputeROC}
\title{ComputeROC}
\usage{
ComputeROC(
  seurat_obj,
  group.by = NULL,
  split_col = NULL,
  features = "hMEs",
  seurat_test = NULL,
  harmony_group_vars = NULL,
  scale_genes = TRUE,
  verbose = FALSE,
  exp_thresh = 0.75,
  return_seurat = TRUE,
  wgcna_name = NULL,
  wgcna_name_test = NULL
)
}
\description{
ComputeROC
}
\examples{
ComputeROC
}
\keyword{scRNA-seq}
