% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWGCNA.R
\name{construct_metacells}
\alias{construct_metacells}
\title{construct_metacells}
\usage{
construct_metacells(
  seurat_obj,
  k = 50,
  name = "agg",
  reduction = "umap",
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{k}{Number of nearest neighbors to aggregate. Default = 50}

\item{name}{A string appended to resulting metalcells. Default = 'agg'}

\item{reduction}{A dimensionality reduction stored in the Seurat object. Default = 'umap'}

\item{assay}{Assay to extract data for aggregation. Default = 'RNA'}

\item{slot}{Slot to extract data for aggregation. Default = 'data'}
}
\description{
This function takes a Seurat object and constructs averaged 'metacells' based
on neighboring cells.
}
\examples{
construct_metacells(pbmc)
}
\keyword{scRNA-seq}
