% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{ConstructMetacells}
\alias{ConstructMetacells}
\title{ConstructMetacells}
\usage{
ConstructMetacells(
  seurat_obj,
  name = "agg",
  ident.group = "seurat_clusters",
  k = 50,
  reduction = "umap",
  assay = "RNA",
  cells.use = NULL,
  slot = "counts",
  meta = NULL,
  return_metacell = FALSE,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{name}{A string appended to resulting metalcells. Default = 'agg'}

\item{k}{Number of nearest neighbors to aggregate. Default = 50}

\item{reduction}{A dimensionality reduction stored in the Seurat object. Default = 'umap'}

\item{assay}{Assay to extract data for aggregation. Default = 'RNA'}

\item{slot}{Slot to extract data for aggregation. Default = 'data'}

\item{return_metacell}{Logical to determine if we return the metacell seurat object (TRUE), or add it to the misc in the original Seurat object (FALSE). Default to FALSE.}
}
\description{
This function takes a Seurat object and constructs averaged 'metacells' based
on neighboring cells.
}
\examples{
ConstructMetacells(pbmc)
}
\keyword{scRNA-seq}
