% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructMetaspots.R
\name{ConstructMetaspots}
\alias{ConstructMetaspots}
\title{ConstructMetaspots}
\usage{
ConstructMetaspots(
  cur_seurat,
  mode = "sum",
  assay = "Spatial",
  slot = "counts"
)
}
\arguments{
\item{cur_seurat}{A Seurat object}

\item{mode}{"sum" or "average"}

\item{assay}{Assay to extract data for aggregation. Default = 'Spatial'}

\item{slot}{Slot to extract data for aggregation. Default = 'counts'}
}
\description{
Computes metaspots in a given Seurat object containing spatial transcriptomics data.
This function is called by MetaspotsByGroups and should NOT be run directly!
}
\examples{

}
\keyword{ST}
