% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructNetwork.R
\name{ConstructNetwork}
\alias{ConstructNetwork}
\title{ConstructNetwork}
\usage{
ConstructNetwork(
  seurat_obj,
  soft_power = NULL,
  min_power = 3,
  tom_outdir = "TOM",
  tom_name = NULL,
  consensus = FALSE,
  overwrite_tom = FALSE,
  wgcna_name = NULL,
  blocks = NULL,
  maxBlockSize = 30000,
  randomSeed = 12345,
  corType = "pearson",
  consensusQuantile = 0.3,
  networkType = "signed",
  TOMType = "signed",
  TOMDenom = "min",
  scaleTOMs = TRUE,
  scaleQuantile = 0.8,
  sampleForScaling = TRUE,
  sampleForScalingFactor = 1000,
  useDiskCache = TRUE,
  chunkSize = NULL,
  deepSplit = 4,
  pamStage = FALSE,
  detectCutHeight = 0.995,
  minModuleSize = 50,
  mergeCutHeight = 0.2,
  saveConsensusTOMs = TRUE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{soft_power}{the soft power used for network construction. Automatically selected by default.}

\item{min_power}{the smallest soft power to be selected if soft_power=NULL}

\item{tom_outdir}{path to the directory where the TOM will be written}

\item{tom_name}{prefix name given to the TOM output file}

\item{consensus}{flag indicating whether or not to perform Consensus network analysis}

\item{wgcna_name}{name of the WGCNA experiment}

\item{...}{additional parameters passed to blockwiseConsensusModules}
}
\value{
seurat_obj with the co-expression network and gene modules computed for the selected wgcna experiment
}
\description{
Constructs a co-expression network and groups genes into modules
given a Seurat object that has been prepared for network analysis.
}
\details{
ConstructNetwork builds a co-expression network and identifies clusters of
highly co-expressed genes (modules) from the metacell or metaspot
expression matrix stored in the Seurat object. Before running this function,
the following functions must be run on the input Seurat object:
\enumerate{
\item SetupForWGCNA
\item MetacellsByGroups or MetaspotsByGroups, and NormalizeMetacells
\item SetDatExpr or SetMultiExpr
\item TestSoftPowers or TestSoftPowersConsensus
}

This function can also be used to perform consensus network analysis if consensus=TRUE
is selected. ConstructNetwork calls the WGCNA function blockwiseConsensusModules
to compute the adjacency matrix, topological overlap matrix, and to run the
Dynamic Tree Cut algorithm to identify gene modules. blockwiseConsensusModules
has numerous parameters but here we have selected default parameters that we
have found to provide reasonable results on a variety of single-cell and
spatial transcriptomic datasets.
}
