% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstructTFNetwork.R
\name{ConstructTFNetwork}
\alias{ConstructTFNetwork}
\title{ConstructTFNetwork}
\usage{
ConstructTFNetwork(seurat_obj, model_params, nfold = 5, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{model_params}{a list of model parameters to pass to xgboost}

\item{nfold}{number of folds for cross validation}

\item{wgcna_name}{name of the WGCNA experiment}
}
\value{
seurat_obj with the TF network added
}
\description{
Construct a network of transcription factors and target genes based on gene co-expression
}
\details{
ConstructTFNetwork uses motif-gene information to build a directed network of transcription
factors (TFs) and target genes. XGBoost regression is leveraged to model the expression of
each gene based on its candidate TF regulators. This analysis gives us information about
how important each TF is for predicting each gene, allowing us to prioritize the most likely
regulators of each gene. This process is done on the gene expression matrix stored with SetDatExpr,
which is typically the hdWGCNA metacell gene expression matrix.
}
