% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{DimPlotMetacells}
\alias{DimPlotMetacells}
\title{DimPlotMetacells}
\usage{
DimPlotMetacells(seurat_obj, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Wrapper function to run Seurat's DimPlot function on the metacell object.
}
\examples{
NormalizeMetadata
}
\keyword{scRNA-seq}
