% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DoHubGeneHeatmap}
\alias{DoHubGeneHeatmap}
\title{Plots gene expression of hub genes as a heatmap}
\usage{
DoHubGeneHeatmap(
  seurat_obj,
  n_hubs = 10,
  n_cells = 200,
  group.by = NULL,
  module_names = NULL,
  combine = TRUE,
  draw.lines = TRUE,
  disp.min = -2.5,
  disp.max = 2.5,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
This function makes an expression heatmap of the top n hub genes per module
using Seurat's DoHeatmap, and then assembles them all into one big heatmap.
}
\examples{
DoHubGeneHeatmap
}
\keyword{scRNA-seq}
