% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{EnrichrBarPlot}
\alias{EnrichrBarPlot}
\title{EnrichrBarPlot}
\usage{
EnrichrBarPlot(
  seurat_obj,
  outdir = "enrichr_plots",
  n_terms = 25,
  plot_size = c(6, 15),
  wgcna_name = NULL,
  logscale = FALSE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}
}
\description{
Makes barplots from Enrichr data
}
\examples{
EnrichrBarPlot
}
\keyword{scRNA-seq}
