% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{EnrichrDotPlot}
\alias{EnrichrDotPlot}
\title{EnrichrDotPlot}
\usage{
EnrichrDotPlot(
  seurat_obj,
  database,
  mods = "all",
  n_terms = 3,
  break_ties = TRUE,
  logscale = TRUE,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{database}{name of the enrichr database to plot.}

\item{mods}{names of modules to plot. All modules are plotted if mods='all' (default)}

\item{n_terms}{number of enriched terms to plot for each module}

\item{break_ties}{logical controlling whether or not to randomly select terms with equal enrichments to precisely enforce n_terms.}

\item{logscale}{logical controlling whether to plot the enrichment on a log scale.}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Makes barplots from Enrichr data
}
\examples{
EnrichrDotPlot
}
\keyword{scRNA-seq}
