% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{FindAllDMEs}
\alias{FindAllDMEs}
\title{FindAllDMEs}
\usage{
FindAllDMEs(
  seurat_obj,
  group.by,
  harmonized = TRUE,
  wgcna_name = NULL,
  add_missing = FALSE,
  test.use = "wilcox",
  only.pos = FALSE,
  logfc.threshold = 0,
  min.pct = 0,
  verbose = FALSE,
  pseudocount.use = 0,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{group.by}{column in seurat_obj@meta.data containing cell grouping information}

\item{harmonized}{logical determining whether or not to use harmonized MEs}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}

\item{add_missing}{logical determining whether or not to add missing modules back into the resulting dataframe with NA values.}

\item{...}{Additional parameters for the FindMarkers function}
}
\value{
A dataframe contaning differential ME results
}
\description{
Modification of the Seurat function FindMarkers used to perform iterative one-versus-all differential module eigengene testing given a group of cells (ie clusters, cell types, etc).
}
\examples{
FindAllDMEs
}
\keyword{scRNA-seq}
