% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regulons.R
\name{FindDifferentialRegulons}
\alias{FindDifferentialRegulons}
\title{FindDifferentialRegulons}
\usage{
FindDifferentialRegulons(
  seurat_obj,
  barcodes1,
  barcodes2,
  assay = "RNA",
  slot = "data",
  layer = "data",
  wgcna_name = NULL,
  test.use = "wilcox",
  only.pos = FALSE,
  logfc.threshold = 0,
  min.pct = 0,
  verbose = FALSE,
  pseudocount.use = 0,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{barcodes1}{character vector containing cell barcodes for the first group to test. Positive fold-change means up-regulated in this group.}

\item{barcodes2}{character vector containing cell barcodes for the second group to test. Negative fold-change means up-regulated in this group.}

\item{assay}{Assay to extract data for aggregation. Default = 'RNA'}

\item{slot}{Slot to extract data for aggregation. Default = 'counts'. Slot is used with Seurat v4 instead of layer.}

\item{layer}{Layer to extract data for aggregation. Default = 'counts'. Layer is used with Seurat v5 instead of slot.}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}

\item{...}{Additional parameters for the Seurat FindMarkers function}
}
\value{
A dataframe contaning differential regulon results
}
\description{
Function to compare TF regulon scores between two sets of cell barcodes.
}
\details{
FindDifferentialRegulons compares two groups based on their TF regulon scores.
Three comparisons are made for different sets of features: positive regulon scores, negative regulon scores, and gene expression.
The same settings for the test will be used for all three tests.
}
