% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMEs.R
\name{FindDMEs}
\alias{FindDMEs}
\title{FindDMEs}
\usage{
FindDMEs(
  seurat_obj,
  barcodes1,
  barcodes2,
  features = "MEs",
  harmonized = TRUE,
  wgcna_name = NULL,
  add_missing = FALSE,
  test.use = "wilcox",
  only.pos = FALSE,
  logfc.threshold = 0,
  min.pct = 0,
  verbose = FALSE,
  pseudocount.use = 0,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{barcodes1}{character vector containing cell barcodes for the first group to test. Positive fold-change means up-regulated in this group.}

\item{barcodes2}{character vector containing cell barcodes for the second group to test. Negative fold-change means up-regulated in this group.}

\item{features}{indicate whether to use "MEs" or "ModuleScores" for the comparison}

\item{harmonized}{logical determining whether or not to use harmonized MEs}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}

\item{add_missing}{logical determining whether or not to add missing modules back into the resulting dataframe with NA values.}

\item{...}{Additional parameters for the FindMarkers function}
}
\value{
A dataframe contaning differential ME results
}
\description{
Function to compare expression levels of co-expression modules between two sets of cell barcodes.
}
\examples{
FindDMEs
}
\keyword{scRNA-seq}
