% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetActiveWGCNAName}
\alias{GetActiveWGCNAName}
\title{GetActiveWGCNAName}
\usage{
GetActiveWGCNAName(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
GetActiveWGCNAName
}
\keyword{scRNA-seq}
