% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetEnrichrTable}
\alias{GetEnrichrTable}
\title{GetEnrichrTable}
\usage{
GetEnrichrTable(seurat_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
GetEnrichrTable
}
\keyword{scRNA-seq}
