% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetHubGenes}
\alias{GetHubGenes}
\title{GetHubGenes}
\usage{
GetHubGenes(seurat_obj, n_hubs = 10, mods = NULL, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{n_hubs}{the number of hub genes to select for each module}

\item{mods}{list of modules, selects all modules by default}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Extract the top N hub genes for a given set of modules. This function outputs
a table with the gene name, the module, and the kME for that module for the
top N hub genes.
}
\keyword{scRNA-seq}
