% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetMEs}
\alias{GetMEs}
\title{GetMEs}
\usage{
GetMEs(seurat_obj, harmonized = TRUE, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{harmonized}{logical indicating whether MEs have been harmonized}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Function to retrieve module eigengens from Seurat object.
}
\keyword{scRNA-seq}
