% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetModulePreservation}
\alias{GetModulePreservation}
\title{GetModulePreservation}
\usage{
GetModulePreservation(seurat_obj, mod_name, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{mod_name}{name of the module preservation test to store}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
GetModulePreservation
}
\keyword{scRNA-seq}
