% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetMotifMatrix}
\alias{GetMotifMatrix}
\title{GetMotifMatrix}
\usage{
GetMotifMatrix(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
GetMotifMatrix
}
\examples{
GetMotifMatrix
}
\keyword{scRNA-seq}
