% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetMotifs}
\alias{GetMotifs}
\title{GetMotifs}
\usage{
GetMotifs(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
GetMotifs
}
\examples{
GetMotifs
}
\keyword{scRNA-seq}
