% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetMotifTargets}
\alias{GetMotifTargets}
\title{GetMotifTargets}
\usage{
GetMotifTargets(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
GetMotifTargets
}
\examples{
GetMotifTargets
}
\keyword{scRNA-seq}
