% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetMultiExpr}
\alias{GetMultiExpr}
\title{GetMultiExpr}
\usage{
GetMultiExpr(seurat_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
This function gets the expression matrix from the metacell object.
}
\examples{
GetDatExpr
}
\keyword{scRNA-seq}
