% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetPFMList}
\alias{GetPFMList}
\title{GetPFMList}
\usage{
GetPFMList(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
GetPFMList
}
\examples{
GetPFMList
}
\keyword{scRNA-seq}
