% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetRegulonScores}
\alias{GetRegulonScores}
\title{GetRegulonScores}
\usage{
GetRegulonScores(seurat_obj, target_type, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{target_type}{dataframe storing the TF regulon scores}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
GetRegulonScores
}
\keyword{scRNA-seq}
