% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{HubGeneNetworkPlot}
\alias{HubGeneNetworkPlot}
\title{HubGeneNetworkPlot}
\usage{
HubGeneNetworkPlot(
  seurat_obj,
  mods = "all",
  n_hubs = 6,
  n_other = 3,
  plot_size = c(6, 6),
  wgcna_name = NULL,
  edge.alpha = 0.25,
  vertex.label.cex = 0.5,
  hub.vertex.size = 4,
  other.vertex.size = 1,
  repulse.exp = 3,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}
}
\description{
Makes barplots from Enrichr data
}
\examples{
HubGeneNetworkPlot
}
\keyword{scRNA-seq}
