% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWGCNA.R
\name{metacells_by_groups}
\alias{metacells_by_groups}
\title{metacells_by_groups}
\usage{
metacells_by_groups(
  seurat_obj,
  group.by = c("seurat_clusters"),
  k = 50,
  reduction = "umap",
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{group.by}{A character vector of Seurat metadata column names representing groups for which metacells will be computed. Default = 'seurat_clusters'}

\item{k}{Number of nearest neighbors to aggregate. Default = 50}

\item{reduction}{A dimensionality reduction stored in the Seurat object. Default = 'umap'}

\item{assay}{Assay to extract data for aggregation. Default = 'RNA'}

\item{slot}{Slot to extract data for aggregation. Default = 'data'}

\item{name}{A string appended to resulting metalcells. Default = 'agg'}
}
\description{
This function takes a Seurat object and constructs averaged 'metacells' based
on neighboring cells in provided groupings, such as cluster or cell type.
}
\examples{
metacells_by_groups(pbmc)
}
\keyword{scRNA-seq}
