% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metaspots.R
\name{MetaspotsByGroups}
\alias{MetaspotsByGroups}
\title{MetaspotsByGroups}
\usage{
MetaspotsByGroups(
  seurat_obj,
  group.by = c("seurat_clusters"),
  ident.group = "seurat_clusters",
  assay = "Spatial",
  slot = "counts",
  mode = "sum",
  min_spots = 50,
  wgcna_name = NULL
)
}
\arguments{
\item{group.by}{A character vector of Seurat metadata column names representing groups for which metacells will be computed.}

\item{assay}{Assay to extract data for aggregation. Default = 'Spatial'}

\item{slot}{Slot to extract data for aggregation. Default = 'counts'}

\item{mode}{determines how to make gene expression profiles for metacells from their constituent single cells. Options are "average" or "sum".}

\item{min_spots}{the minimum number of spots in a particular grouping to construct metaspots}

\item{wgcna_name}{name of the WGCNA experiment}

\item{cur_seurat}{A Seurat object}
}
\description{
Computes metaspots in a given Seurat object containing spatial transcriptomics data.
}
