% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{ModuleCorrelogram}
\alias{ModuleCorrelogram}
\title{ModuleCorrelogram}
\usage{
ModuleCorrelogram(
  seurat_obj,
  MEs2 = NULL,
  features = "hMEs",
  order = "original",
  method = "ellipse",
  exclude_grey = TRUE,
  type = "upper",
  tl.col = "black",
  tl.srt = 45,
  sig.level = c(1e-04, 0.001, 0.01, 0.05),
  pch.cex = 0.7,
  col = NULL,
  ncolors = 200,
  wgcna_name = NULL,
  wgcna_name2 = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{features}{What to plot? Can select hMEs, MEs, scores, or average}
}
\description{
Plot Module Eigengene correlogram
}
\examples{
MECorrelogram
}
\keyword{scRNA-seq}
