% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{ModuleCorrNetwork}
\alias{ModuleCorrNetwork}
\title{ModuleCorrNetworks}
\usage{
ModuleCorrNetwork(
  seurat_obj,
  wgcna_name = NULL,
  cluster_col = NULL,
  exclude_grey = TRUE,
  features = "hMEs",
  reduction = "umap",
  cor_cutoff = 0.2,
  label_vertices = FALSE,
  edge_scale = 5,
  vertex_size = 15,
  niter = 100,
  vertex_frame = FALSE
)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Plot Module Eigengene correlogram
}
\examples{
ModuleCorrNetwork
}
\keyword{scRNA-seq}
