% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ModuleEigengenes}
\alias{ModuleEigengenes}
\title{ModuleEigengenes}
\usage{
ModuleEigengenes(
  seurat_obj,
  group.by.vars = NULL,
  modules = NULL,
  verbose = TRUE,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Computes module eigengenes for all WGCNA co-expression modules
}
\examples{
 ModuleEigengenes(pbmc)
}
\keyword{scRNA-seq}
