% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ModuleExprScore}
\alias{ModuleExprScore}
\title{ModuleExprScores}
\usage{
ModuleExprScore(
  seurat_obj,
  n_genes = 25,
  wgcna_name = NULL,
  method = "Seurat",
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{method}{Seurat or UCell?}
}
\description{
Computes module eigengenes for all WGCNA co-expression modules
}
\examples{
 ModuleExprScores (pbmc)
}
\keyword{scRNA-seq}
