% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ModuleFeaturePlot}
\alias{ModuleFeaturePlot}
\title{ModuleFeaturePlot}
\usage{
ModuleFeaturePlot(
  seurat_obj,
  module_names = NULL,
  wgcna_name = NULL,
  reduction = "umap",
  features = "hMEs",
  order_points = TRUE,
  restrict_range = TRUE,
  point_size = 0.5,
  alpha = 1,
  label_legend = FALSE,
  ucell = FALSE,
  raster = FALSE,
  raster_dpi = 500,
  raster_scale = 1,
  plot_ratio = 1,
  title = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{features}{What to plot? Can select hMEs, MEs, scores, or average}

\item{order}{TRUE, FALSE, or "shuffle" are valid options}
}
\description{
Plot module eigengenes as a FeaturePlot
}
\examples{
ModuleFeaturePlot
}
\keyword{scRNA-seq}
