% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ModuleNetworkPlot}
\alias{ModuleNetworkPlot}
\title{ModuleNetworkPlot}
\usage{
ModuleNetworkPlot(
  seurat_obj,
  mods = "all",
  outdir = "ModuleNetworks",
  plot_size = c(6, 6),
  wgcna_name = NULL,
  edge.alpha = 0.25,
  vertex.label.cex = 1,
  vertex.size = 6,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}
}
\description{
Makes barplots from Enrichr data
}
\examples{
ModuleNetworkPlot
}
\keyword{scRNA-seq}
