% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ModulePreservation}
\alias{ModulePreservation}
\title{ModulePreservation}
\usage{
ModulePreservation(
  seurat_obj,
  seurat_ref,
  name,
  n_permutations = 500,
  parallel = FALSE,
  seed = 12345,
  return_raw = FALSE,
  wgcna_name = NULL,
  wgcna_name_ref = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{seurat_ref}{A Seurat object serving as the reference for the module preservation analysis}

\item{name}{The name to give the module preservation analysis.}

\item{n_permutations}{Number of permutations for the module preservation test.}

\item{parallel}{logical determining whether to run preservation analysis in parallel}

\item{seed}{random seed for the permutation analysis.}

\item{return_raw}{if TRUE, returns the module preservation statistics, else returns seurat_obj with the stats added to the scWGCNA experiment.}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}

\item{wgcna_name_ref}{The name of the scWGCNA experiment in the seurat_ref@misc slot}
}
\description{
Computes module preservation statistics in a query dataset for a given reference dataset
}
\examples{
ModulePreservation
}
\keyword{scRNA-seq}
