% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleRegulatoryNetwork.R
\name{ModuleRegulatoryNetwork}
\alias{ModuleRegulatoryNetwork}
\title{ModuleRegulatoryNetwork}
\usage{
ModuleRegulatoryNetwork(seurat_obj, TFs_only = TRUE, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the single-cell data and WGCNA results.}

\item{TFs_only}{Logical; if TRUE (default), only transcription factor (TF) genes are
included in the regulatory network. If FALSE, the network includes all genes.}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\value{
A data frame with the following columns:
\item{source}{The source module in the regulatory interaction.}
\item{target}{The target module in the regulatory interaction.}
\item{n_pos}{The number of positive regulatory links from the source module to the target module.}
\item{sum_pos}{The sum of the gain values for the positive links.}
\item{n_neg}{The number of negative regulatory links from the source module to the target module.}
\item{sum_neg}{The sum of the gain values for the negative links.}
\item{mean_pos}{The average gain of the positive links.}
\item{mean_neg}{The average gain of the negative links.}
\item{score_pos}{The number of positive links normalized by the number of TFs in the source module.}
\item{score_neg}{The number of negative links normalized by the number of TFs in the source module.}
}
\description{
Summarizes Transcription Factor regulatory networks across co-expression modules.
}
\details{
This function summarizes transcription factor regulatory networks across modules to infer module-module relationships. The number of directed TF links across modules are counted and normalized. The strength of these links from the XGBoost model are also tracked.
}
