% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ModuleUMAPPlot}
\alias{ModuleUMAPPlot}
\title{ModuleUMAPPlot}
\usage{
ModuleUMAPPlot(
  seurat_obj,
  sample_edges = TRUE,
  edge_prop = 0.2,
  label_hubs = 5,
  edge.alpha = 0.25,
  vertex.label.cex = 0.5,
  hub.vertex.size = 4,
  other.vertex.size = 1,
  return_graph = FALSE,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Makes a igraph network plot using the module UMAP
}
\examples{
ModuleUMAPPlot
}
\keyword{scRNA-seq}
