% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{MotifOverlapBarPlot}
\alias{MotifOverlapBarPlot}
\title{Displays the top n TFs in a set of modules as a bar plot}
\usage{
MotifOverlapBarPlot(
  seurat_obj,
  n_tfs = 10,
  plot_size = c(5, 6),
  outdir = "MotifOverlaps/",
  motif_font = "helvetica_regular",
  module_names = NULL,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Displays the top n TFs in a set of modules as a bar plot
}
\examples{
MotifOverlapBarPlot
}
\keyword{scRNA-seq}
