% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motifs.R
\name{MotifScan}
\alias{MotifScan}
\title{Scan gene promoters for a set of TF PWMs}
\usage{
MotifScan(seurat_obj, species_genome, pfm, EnsDb, wgcna_name = NULL)
}
\description{
Scan gene promoters for a set of TF PWMs
}
\examples{
MotifScan
}
\keyword{scRNA-seq}
