% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{MotifTargetScore}
\alias{MotifTargetScore}
\title{MotifTargetScore}
\usage{
MotifTargetScore(
  seurat_obj,
  method = "Seurat",
  wgcna_genes = TRUE,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{method}{Seurat or UCell?}
}
\description{
Computes gene expression scores for TF Motif target genes based on the MotifScan.
}
\examples{
MotifTargetScore(pbmc)
}
\keyword{scRNA-seq}
