% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{OverlapBarPlot}
\alias{OverlapBarPlot}
\title{Plots the results from OverlapModulesDEGs as a bar plot}
\usage{
OverlapBarPlot(
  overlap_df,
  plot_var = "odds_ratio",
  logscale = TRUE,
  neglog = FALSE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Plots the results from OverlapModulesDEGs as a bar plot
}
\examples{
OverlapBarPlot
}
\keyword{scRNA-seq}
