% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{OverlapDotPlot}
\alias{OverlapDotPlot}
\title{OverlapDotPlot}
\usage{
OverlapDotPlot(
  overlap_df,
  plot_var = "odds_ratio",
  logscale = TRUE,
  neglog = FALSE,
  plot_significance = TRUE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Makes barplots from Enrichr data
}
\examples{
OverlapDotPlot
}
\keyword{scRNA-seq}
