% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{OverlapModulesDEGs}
\alias{OverlapModulesDEGs}
\title{OverlapModulesDEGs}
\usage{
OverlapModulesDEGs(seurat_obj, deg_df, wgcna_name = NULL, fc_cutoff = 0.5, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}
}
\description{
Computes intramodular connectivity (kME) based on module eigengenes.
}
\examples{
OverlapModulesDEGs
}
\keyword{scRNA-seq}
