% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motifs.R
\name{OverlapModulesMotifs}
\alias{OverlapModulesMotifs}
\title{Overlap modules with TF target genes}
\usage{
OverlapModulesMotifs(seurat_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Overlap modules with TF target genes
}
\examples{
OverlapModulesDEGs
}
\keyword{scRNA-seq}
