% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotDendrogram}
\alias{PlotDendrogram}
\title{PlotDendrogram}
\usage{
PlotDendrogram(
  seurat_obj,
  groupLabels = "Module colors",
  wgcna_name = NULL,
  dendroLabels = FALSE,
  hang = 0.03,
  addGuide = TRUE,
  guideHang = 0.05,
  main = "",
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Plot WGCNA dendrogram
}
\examples{
PlotDendrogram
}
\keyword{scRNA-seq}
