% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotDMEsLollipop}
\alias{PlotDMEsLollipop}
\title{PlotDMEsLollipop}
\usage{
PlotDMEsLollipop(
  seurat_obj,
  DMEs,
  wgcna_name,
  group.by = NULL,
  comparison = NULL,
  pvalue,
  avg_log2FC = "avg_log2FC"
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{DMEs}{dataframe output from FindDMEs or FindAllDMEs}

\item{group.by}{the column name of the selected comparison in the DMEs dataframe}

\item{comparison}{character vector or a list of character vectors containing the comparison in the DMEs dataframe}

\item{pvalue}{p_value or fdr used for the comparison in the DMEs dataframe}
}
\value{
A ggplot object
PlotDMEsLollipop
}
\description{
Plotting function for the results of FindDMEs and FindAllDMEs
}
\keyword{scRNA-seq}
