% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotKMEs}
\alias{PlotKMEs}
\title{PlotKMEs}
\usage{
PlotKMEs(
  seurat_obj,
  n_hubs = 10,
  text_size = 2,
  ncol = 5,
  plot_widths = c(3, 2),
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{n_hubs}{number of hub genes to display}

\item{text_size}{controls the size of the hub gene text}

\item{ncol}{number of columns to display individual plots}

\item{plot_widths}{the relative width between the kME rank plot and the hub gene text}

\item{wgcna_name}{the name of the WGCNA experiment in the seurat object}
}
\description{
Plotting function to show genes by kME value in each module
}
\keyword{scRNA-seq}
