% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotLollipop}
\alias{PlotLollipop}
\title{PlotLollipop}
\usage{
PlotLollipop(modules, cur_DMEs, pvalue, avg_log2FC = "avg_log2FC")
}
\arguments{
\item{pvalue}{p_value or fdr used for the comparison in the DMEs dataframe}

\item{seurat_obj}{A Seurat object}

\item{DMEs}{dataframe output from FindDMEs or FindAllDMEs}

\item{group.by}{the column name of the selected comparison in the DMEs dataframe}

\item{comparison}{character vector or a list of character vectors containing the comparison in the DMEs dataframe}
}
\value{
A ggplot object
}
\description{
An Internal function for PlotDMEsLollipop to Plotting function for the results of FindDMEs and FindAllDMEs
}
\examples{
plot_list <- PlotLollipop(modules, DMEs, cur_title= c("Group1_vs_Group2"), pvalue, avg_log2FC = 'avg_log2FC')
}
\keyword{scRNA-seq}
