% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotModulePreservation}
\alias{PlotModulePreservation}
\title{PlotModulePreservation}
\usage{
PlotModulePreservation(
  seurat_obj,
  name,
  statistics = "summary",
  plot_labels = TRUE,
  label_size = 4,
  mod_point_size = 4,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{name}{The name of the module preservation analysis to plot given in ModulePreservation}

\item{statistics}{Which module preservation statistics to plot? Choices are summary, all, or a custom list}

\item{plot_labels}{logical determining whether to plot the module labels}

\item{label_size}{the size of the module labels}

\item{mod_point_size}{the size of the points in each plot}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Plotting function for Module Preservation statistics
}
\details{
This function creates a scatter plot showing the module preservation statistics for each module
compared to the size of the module (number of genes).
}
