% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotModulePreservationLollipop}
\alias{PlotModulePreservationLollipop}
\title{PlotModulePreservationLollipop}
\usage{
PlotModulePreservationLollipop(
  seurat_obj,
  name,
  features = NULL,
  fdr = TRUE,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{name}{The name to give the module preservation analysis.}

\item{features}{The name of the module preservation features to plot.}

\item{fdr}{logical indicating whether or not to plot FDR-corrected p-values}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\value{
ggplot object containing the PlotModulePreservationLollipop
}
\description{
Plots a ranked lollipop plot of co-expression modules based on the results of module preservation analysis.
}
\details{
PlotModulePreservationLollipop generates a lollipop plot showing module preservation results. If the module
preservation test was performed using the WGCNA method, the statistic that will be shown is the Z-summary
preservation statistic. If the analysis was performed using NetRep, then the statistic that will be shown is
the FDR corrected averaged p-values from the module preservation permutation test.
}
