% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{PlotModuleTrajectory}
\alias{PlotModuleTrajectory}
\title{PlotModuleTrajectory}
\usage{
PlotModuleTrajectory(
  seurat_obj,
  pseudotime_col = "pseudotime",
  n_bins = 20,
  harmonized = TRUE,
  ncol = 4,
  point_size = 1,
  line_size = 1,
  se = TRUE,
  group_colors = NULL,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{pseudotime_col}{name of the column in seurat_obj@meta.data containing pseudotime information. Multiple names can be passed to plot multiple trajectories simultaneously.}

\item{n_bins}{number of pseudotime bins/windows used to smooth the MEs}

\item{harmonized}{logical determining whether or not to use the harmonized MEs}

\item{point_size}{size of the points in the plot}

\item{line_size}{width of the lines in the plot}

\item{se}{logical determining whether or not to show the standard error on the plot}

\item{group_colors}{optional list of colors to differentiate multiple pseudotime trajectories. Must be the same length as pseudotime_col}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}

\item{n_col}{number of columns for different modules in the plot}
}
\description{
Plots module eigengene dynamics over a pseudotime trajectory.
}
