% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotSoftPowers}
\alias{PlotSoftPowers}
\title{PlotSoftPowers}
\usage{
PlotSoftPowers(
  seurat_obj,
  selected_power = NULL,
  point_size = 5,
  text_size = 3,
  plot_connectivity = TRUE,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{selected_power}{power to highlight in the plots}

\item{point_size}{the size of the points in the plot}

\item{text_size}{the size of the text in the plot}

\item{plot_connectivity}{logical indicating whether to plot the connectivity in addition to the scale free topplogy fit.}

\item{wgcna_name}{The name of the WGCNA experiment in seurat_obj}
}
\description{
Plot Soft Power Threshold results
}
\examples{
PlotSoftPowers
}
\keyword{scRNA-seq}
