% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regulons.R
\name{RegulonScores}
\alias{RegulonScores}
\title{RegulonScores}
\usage{
RegulonScores(
  seurat_obj,
  target_type = "positive",
  cor_thresh = 0.05,
  exclude_grey_genes = TRUE,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{target_type}{Which types of TF target genes to compute scores for? "positive", "negative", or "both"}

\item{cor_thresh}{threshold for TF-gene correlation for genes to be included in the regulon score}

\item{exclude_grey_genes}{option to exclude genes that are in the grey module from the regulon scores}

\item{wgcna_name}{name of the WGCNA experiment}
}
\value{
seurat_obj with the TF Regulon Scores added
}
\description{
Calculate expression scores for TF Regulons
}
\details{
RegulonScores calculates expression signatures for each TF regulon using the UCell algorithm
This function can calculate separate scores for TF regulons for target genes that are positively
or negatively correlated with the TF, representing putative acivated or repressed genes. These scores
conveniently summarize the expression levels of the entire TF regulon, similar to module eigengenes
for the co-expression network analyssis.
}
