% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReassignModules.R
\name{ResetModuleColors}
\alias{ResetModuleColors}
\title{ResetModuleColors}
\usage{
ResetModuleColors(seurat_obj, new_colors, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{new_colors}{a character vector containing the new colors}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Reset the unique color for each hdWGCNA module
}
\keyword{scRNA-seq}
