% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{RunEnrichr}
\alias{RunEnrichr}
\title{RunEnrichr}
\usage{
RunEnrichr(
  seurat_obj,
  dbs = c("GO_Biological_Process_2021", "GO_Cellular_Component_2021",
    "GO_Molecular_Function_2021"),
  max_genes = 100,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}
}
\description{
Computes intramodular connectivity (kME) based on module eigengenes.
}
\examples{
RunEnrichr
}
\keyword{scRNA-seq}
