% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{RunHarmonyMetacells}
\alias{RunHarmonyMetacells}
\title{RunHarmonyMetacells}
\usage{
RunHarmonyMetacells(seurat_obj, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Wrapper function to run harmony's RunHarmony function on the metacell object.
}
