% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{RunModuleUMAP}
\alias{RunModuleUMAP}
\title{Run UMAP on co-expression matrix using hub genes as features.}
\usage{
RunModuleUMAP(
  seurat_obj,
  n_hubs = 50,
  exclude_grey = TRUE,
  harmonized = TRUE,
  wgcna_name = NULL,
  n_neighbors = 25,
  metric = "cosine",
  spread = 1,
  min_dist = 0.4,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Run UMAP on co-expression matrix using hub genes as features.
}
\examples{
RunModuleUMAP
}
\keyword{scRNA-seq}
