% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_wrappers.R
\name{RunPCAMetacells}
\alias{RunPCAMetacells}
\title{RunPCAMetacells}
\usage{
RunPCAMetacells(seurat_obj, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Wrapper function to run Seurat's RunPCA function on the metacell object.
}
\examples{
NormalizeMetadata
}
\keyword{scRNA-seq}
