% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{ScaleMetacells}
\alias{ScaleMetacells}
\title{ScaleMetacells}
\usage{
ScaleMetacells(seurat_obj, wgcna_name = NULL, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Wrapper function to run Seurat's ScaleData function on the metacell object.
}
