% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectNetworkGenes.R, R/SetupForWGCNA.R
\name{SelectNetworkGenes}
\alias{SelectNetworkGenes}
\title{SelectNetworkGenes}
\usage{
SelectNetworkGenes(
  seurat_obj,
  gene_select = "variable",
  fraction = 0.05,
  group.by = NULL,
  gene_list = NULL,
  wgcna_name = NULL
)

SelectNetworkGenes(
  seurat_obj,
  gene_select = "variable",
  fraction = 0.05,
  group.by = NULL,
  gene_list = NULL,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{fraction}{A numeric that determines the minimum cells that a gene must be expressed in order to be included. For example, fraction = 0.05 means that 5\% of cells must express a gene (count > 0) for it to be included.}

\item{gene_list}{A character string of gene names, only used if type = "custom"}

\item{type}{How to select genes? Select "variable", "fraction", "all", or "custom".}
}
\description{
This function
on neighboring cells in provided groupings, such as cluster or cell type.

This function
on neighboring cells in provided groupings, such as cluster or cell type.
}
\examples{
SelectNetworkGenes(pbmc)
SelectNetworkGenes(pbmc)
}
\keyword{scRNA-seq}
