% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetActiveWGCNA}
\alias{SetActiveWGCNA}
\title{SetActiveWGCNA}
\usage{
SetActiveWGCNA(seurat_obj, wgcna_name)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetActiveWGCNA
}
\examples{
SetActiveWGCNA
}
\keyword{scRNA-seq}
