% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetAvgModuleExpr}
\alias{SetAvgModuleExpr}
\title{SetAvgModuleExpr}
\usage{
SetAvgModuleExpr(seurat_obj, avg_mods, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{avg_mods}{dataframe storing the average expression of all genes in the same module}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetAvgModuleExpr
}
