% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetDegrees}
\alias{SetDegrees}
\title{SetDegrees}
\usage{
SetDegrees(seurat_obj, degree_df, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{degree_df}{dataframe containing gene module assignments}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetDegrees
}
