% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetEnrichrRegulonTable}
\alias{SetEnrichrRegulonTable}
\title{SetEnrichRegulonTable}
\usage{
SetEnrichrRegulonTable(seurat_obj, enrich_table, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{enrich_table}{dataframe storing the results of running enrichr}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetEnrichRegulonTable
}
\keyword{scRNA-seq}
