% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMELoadings}
\alias{SetMELoadings}
\title{SetMELoadings}
\usage{
SetMELoadings(seurat_obj, loadings, harmonized = TRUE, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{loadings}{named numeric vector with eigengene loadings}

\item{harmonized}{logical indicating whether MEs have been harmonized}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetMELoadings
}
\keyword{scRNA-seq}
