% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMEs}
\alias{SetMEs}
\title{SetMEs}
\usage{
SetMEs(seurat_obj, MEs, harmonized = TRUE, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{MEs}{dataframe or matrix containing module eigengenes}

\item{harmonized}{logical indicating whether MEs have been harmonized}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetMEs
}
\keyword{scRNA-seq}
