% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMetacellObject}
\alias{SetMetacellObject}
\title{SetMetacellObject}
\usage{
SetMetacellObject(seurat_obj, metacell_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{metacell_obj}{metacell Seurat object}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetMetacellObject
}
\keyword{scRNA-seq}
