% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMetacellParams}
\alias{SetMetacellParams}
\title{SetMetacellParams}
\usage{
SetMetacellParams(seurat_obj, params, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{params}{list of WGCNA parameters}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetMetacellParams
}
\keyword{scRNA-seq}
