% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetModulePreservation}
\alias{SetModulePreservation}
\title{SetModulePreservation}
\usage{
SetModulePreservation(seurat_obj, mod_pres, mod_name, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{mod_name}{name of the module preservation test to store}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}

\item{mt_cor}{matrix of module-trait correlation results}
}
\description{
SetModulePreservation
}
\keyword{scRNA-seq}
