% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetModuleScores}
\alias{SetModuleScores}
\title{SetModuleScores}
\usage{
SetModuleScores(seurat_obj, mod_scores, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{mod_scores}{dataframe storing the module expression scores}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetModuleScores
}
