% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetModuleTraitCorrelation}
\alias{SetModuleTraitCorrelation}
\title{SetModuleTraitCorrelation}
\usage{
SetModuleTraitCorrelation(seurat_obj, mt_cor, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{mt_cor}{matrix of module-trait correlation results}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetModuleTraitCorrelation
}
\keyword{scRNA-seq}
