% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetModuleUMAP}
\alias{SetModuleUMAP}
\title{SetModuleUMAP}
\usage{
SetModuleUMAP(seurat_obj, umap_df, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{umap_df}{dataframe of UMAP coordinates}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetModuleUMAP
}
\keyword{scRNA-seq}
