% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMotifMatrix}
\alias{SetMotifMatrix}
\title{SetMotifMatrix}
\usage{
SetMotifMatrix(seurat_obj, tf_match)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{tf_match}{matrix containing tf-promoter matches}
}
\description{
SetMotifMatrix
}
\examples{
SetMotifMatrix
}
\keyword{scRNA-seq}
