% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMotifOverlap}
\alias{SetMotifOverlap}
\title{SetMotifOverlap}
\usage{
SetMotifOverlap(seurat_obj, overlap_df, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{overlap_df}{dataframe containing motif-module overlap info}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetMotifOverlap
}
\keyword{scRNA-seq}
