% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMotifs}
\alias{SetMotifs}
\title{SetMotifs}
\usage{
SetMotifs(seurat_obj, motif_df)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{motif_df}{dataframe containing info about the motifs being analyzed}
}
\description{
SetMotifs
}
\examples{
SetMotifs
}
\keyword{scRNA-seq}
