% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMotifScores}
\alias{SetMotifScores}
\title{SetMotifScores}
\usage{
SetMotifScores(seurat_obj, tf_scores, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{tf_scores}{dataframe of tf motif target scores}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetMotifScores
}
\examples{
SetMotifScores
}
\keyword{scRNA-seq}
