% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMotifTargets}
\alias{SetMotifTargets}
\title{SetMotifTargets}
\usage{
SetMotifTargets(seurat_obj, motif_targets)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{motif_targets}{list of motifs and their target genes}
}
\description{
SetMotifTargets
}
\examples{
SetMotifTargets
}
\keyword{scRNA-seq}
