% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetNetworkData}
\alias{SetNetworkData}
\title{SetNetworkData}
\usage{
SetNetworkData(seurat_obj, net, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{net}{list of network data from WGCNA}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetNetworkData
}
\keyword{scRNA-seq}
