% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetPFMList}
\alias{SetPFMList}
\title{SetPFMList}
\usage{
SetPFMList(seurat_obj, pfm_list)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{pfm_list}{list of pfm objects}
}
\description{
SetPFMList
}
\examples{
SetPFMList
}
\keyword{scRNA-seq}
