% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetPowerTable}
\alias{SetPowerTable}
\title{SetPowerTable}
\usage{
SetPowerTable(seurat_obj, power_table, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{power_table}{a dataframe containing the results of the soft power test}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetPowerTable
}
\keyword{scRNA-seq}
