% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetRegulonScores}
\alias{SetRegulonScores}
\title{SetRegulonScores}
\usage{
SetRegulonScores(seurat_obj, regulon_scores, target_type, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{regulon_scores}{dataframe storing the TF regulon scores}

\item{target_type}{dataframe storing the TF regulon scores}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetRegulonScores
}
\keyword{scRNA-seq}
