% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetROCData}
\alias{SetROCData}
\title{SetROCData}
\usage{
SetROCData(seurat_obj, roc_info, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{roc_info}{dataframe storing the ROC results}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetROCData
}
\examples{
SetROCData
}
\keyword{scRNA-seq}
