% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetTFEval}
\alias{SetTFEval}
\title{SetTFEval}
\usage{
SetTFEval(seurat_obj, tf_eval, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{tf_eval}{dataframe storing the TF network evaluation info from ConstructTFNetwork}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetTFEval
}
\keyword{scRNA-seq}
