% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetTFNetwork}
\alias{SetTFNetwork}
\title{SetTFNetwork}
\usage{
SetTFNetwork(seurat_obj, tf_net, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{tf_net}{dataframe storing the TF network info in ConstructTFNetwork}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetTFNetwork
}
\keyword{scRNA-seq}
