% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetTFRegulons}
\alias{SetTFRegulons}
\title{SetTFRegulons}
\usage{
SetTFRegulons(seurat_obj, tf_regulons, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{tf_regulons}{dataframe storing the TF regulon info from AssignTFRegulons}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetTFRegulons
}
\keyword{scRNA-seq}
