% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{SetupForWGCNA}
\alias{SetupForWGCNA}
\title{SetupForWGCNA}
\usage{
SetupForWGCNA(seurat_obj, wgcna_name, group = NULL, features = NULL, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
This function gets the expression matrix from the metacell object.
}
\examples{
SetupForWGCNA(pbmc)
}
\keyword{scRNA-seq}
