% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetWGCNAGenes}
\alias{SetWGCNAGenes}
\title{SetWGCNAGenes}
\usage{
SetWGCNAGenes(seurat_obj, gene_list, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{gene_list}{vector of genes to be used for WGCNA}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetWGCNAGenes
}
\examples{
SetWGCNAGenes
}
\keyword{scRNA-seq}
