% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetWGCNAGroup}
\alias{SetWGCNAGroup}
\title{SetWGCNAGroup}
\usage{
SetWGCNAGroup(seurat_obj, group, wgcna_name)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{group}{group name}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetWGCNAGroup
}
\examples{
SetWGCNAGroup
}
\keyword{scRNA-seq}
