% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{TestSoftPowersConsensus}
\alias{TestSoftPowersConsensus}
\title{TestSoftPowersConsensus}
\usage{
TestSoftPowersConsensus(
  seurat_obj,
  powers = c(seq(1, 10, by = 1), seq(12, 30, by = 2)),
  setDatExpr = TRUE,
  use_metacells = TRUE,
  group.by = NULL,
  group_name = NULL,
  multi.group.by = NULL,
  multi_groups = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{powers}{numeric vector specifying soft powers to test}

\item{setDatExpr}{logical flag indicating whether to run setDatExpr}

\item{...}{additional parameters passed to SetDatExpr}
}
\description{
Compute the scale-free topology model fit for different soft power thresholds separately for each input dataset
}
\examples{
TestSoftPowers(pbmc)
}
\keyword{scRNA-seq}
