% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectModules.R
\name{TransferModuleGenome}
\alias{TransferModuleGenome}
\title{TransferModuleGenome}
\usage{
TransferModuleGenome(
  modules,
  gene_mapping,
  genome1_col = NULL,
  genome2_col = NULL
)
}
\arguments{
\item{modules}{module table from GetModules function}

\item{gene_mapping}{table that has the gene names for both genomes}

\item{genome1_col}{the column in gene_mapping that has gene names for the genome currently present in modules}

\item{genome2_col}{the column in gene_mapping that has gene names for the genome to transfer to}
}
\description{
Takes a module table and a gene mapping table (like from biomart) with gene
names from two genomes in order to switch
}
\examples{
TransferModuleGenome
}
\keyword{scRNA-seq}
