% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleEigengenes.R
\name{ComputeModuleEigengene}
\alias{ComputeModuleEigengene}
\title{ComputeModuleEigengene}
\usage{
ComputeModuleEigengene(
  seurat_obj,
  cur_mod,
  modules,
  group.by.vars = NULL,
  verbose = TRUE,
  vars.to.regress = NULL,
  scale.model.use = "linear",
  pc_dim = 1,
  assay = NULL,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{cur_mod}{name of a module found in seurat_obj@misc[\link{seurat_obj@misc$active_wgcna}]$wgcna_net$colors}

\item{modules}{table containing module / gene assignments, as in GetModules(seurat_obj).}

\item{group.by.vars}{groups to harmonize by}

\item{verbose}{logical indicating whether to print messages}

\item{vars.to.regress}{character vector of variables in seurat_obj@meta.data to regress when running ScaleData}

\item{scale.model.use}{model to scale data when running ScaleData choices are "linear", "poisson", or "negbinom"}

\item{pc_dim}{Which PC to use as the module eigengene? Default to 1.}

\item{assay}{Assay in seurat_obj to compute module eigengenes. Default is DefaultAssay(seurat_obj)}

\item{wgcna_name}{name of the WGCNA experiment}
}
\description{
Internal helper function that computes module eigengene for a single module.
}
