% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ConstructNetwork}
\alias{ConstructNetwork}
\title{ConstructNetwork}
\usage{
ConstructNetwork(
  seurat_obj,
  soft_power = NULL,
  use_metacells = TRUE,
  group.by = NULL,
  group_name = NULL,
  tom_outdir = "TOM",
  blocks = NULL,
  maxBlockSize = 30000,
  randomSeed = 12345,
  corType = "pearson",
  consensusQuantile = 0.3,
  networkType = "signed",
  TOMType = "unsigned",
  TOMDenom = "min",
  scaleTOMs = TRUE,
  scaleQuantile = 0.8,
  sampleForScaling = TRUE,
  sampleForScalingFactor = 1000,
  useDiskCache = TRUE,
  chunkSize = NULL,
  deepSplit = 4,
  pamStage = FALSE,
  detectCutHeight = 0.995,
  minModuleSize = 50,
  mergeCutHeight = 0.2,
  saveConsensusTOMs = TRUE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
This function constructs a co-expression network from a Seurat object
}
\examples{
ConstructNetwork(pbmc)
}
\keyword{scRNA-seq}
