% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{EnrichrBarPlot}
\alias{EnrichrBarPlot}
\title{EnrichrBarPlot}
\usage{
EnrichrBarPlot(
  seurat_obj,
  outdir = "enrichr_plots",
  n_terms = 25,
  plot_size = c(6, 15),
  logscale = FALSE,
  plot_bar_color = NULL,
  plot_text_color = NULL,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{outdir}{directory to place output .pdf files}

\item{n_terms}{the number of terms to plot in each barplot}

\item{plot_size}{the size of the output .pdf files (width, height)}

\item{logscale}{logical controlling whether to plot the enrichment on a log scale}

\item{plot_bar_color}{The color of the bars in the bar plots. Default option (NULL) makes the bars colored using the module's assigned color.}

\item{plot_text_color}{The color of the text labels on the bar plots}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Makes barplots from RunEnrichr output.
}
\examples{
EnrichrBarPlot
}
\keyword{scRNA-seq}
