% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{EnrichrDotPlot}
\alias{EnrichrDotPlot}
\title{EnrichrDotPlot}
\usage{
EnrichrDotPlot(
  seurat_obj,
  database,
  mods = "all",
  outdir = "enrichr_plots",
  n_terms = 3,
  break_ties = TRUE,
  wgcna_name = NULL,
  logscale = TRUE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}
}
\description{
Makes barplots from Enrichr data
}
\examples{
EnrichrBarPlot
}
\keyword{scRNA-seq}
