% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{GetDatExpr}
\alias{GetDatExpr}
\title{GetDatExpr}
\usage{
GetDatExpr(seurat_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
This function gets the WGCNA expression matrix.
}
\examples{
GetDatExpr
}
\keyword{scRNA-seq}
