% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{MetacellsByGroups}
\alias{MetacellsByGroups}
\title{MetacellsByGroups}
\usage{
MetacellsByGroups(
  seurat_obj,
  group.by = c("seurat_clusters"),
  ident.group = "seurat_clusters",
  k = 25,
  reduction = "pca",
  dims = NULL,
  assay = NULL,
  slot = "counts",
  layer = "counts",
  mode = "average",
  cells.use = NULL,
  min_cells = 100,
  max_shared = 15,
  target_metacells = 1000,
  max_iter = 5000,
  verbose = FALSE,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{group.by}{A character vector of Seurat metadata column names representing groups for which metacells will be computed.}

\item{k}{Number of nearest neighbors to aggregate. Default = 50}

\item{reduction}{A dimensionality reduction stored in the Seurat object. Default = 'pca'}

\item{dims}{A vector represnting the dimensions of the reduction to use. Either specify the names of the dimensions or the indices. Default = NULL to include all dims.}

\item{assay}{Assay to extract data for aggregation. Default = 'RNA'}

\item{slot}{Slot to extract data for aggregation. Default = 'counts'. Slot is used with Seurat v4 instead of layer.}

\item{layer}{Layer to extract data for aggregation. Default = 'counts'. Layer is used with Seurat v5 instead of slot.}

\item{mode}{determines how to make gene expression profiles for metacells from their constituent single cells. Options are "average" or "sum".}

\item{min_cells}{the minimum number of cells in a particular grouping to construct metacells}

\item{max_shared}{the maximum number of cells to be shared across two metacells}

\item{target_metacells}{the maximum target number of metacells to construct}

\item{max_iter}{the maximum number of iterations in the metacells bootstrapping loop}

\item{verbose}{logical indicating whether to print additional information}

\item{wgcna_name}{name of the WGCNA experiment}

\item{name}{A string appended to resulting metalcells. Default = 'agg'}
}
\value{
seurat_obj with a metacell seurat_obj stored in the specified WGCNA experiment
}
\description{
This function takes a Seurat object and constructs averaged 'metacells' based
on neighboring cells in provided groupings, such as cluster or cell type.
}
\details{
MetacellsByGroups merges transcriptomically similar cells into "metacells".
Given a dimensionally-reduced representation of the input dataset, this algorithm
first uses KNN to identify similar cells. A bootstrapped sampling procedure is then
used to group together similar cells until convergence is reached. Importantly,
this procedure is done in a context-specific manner based on the provided group.by parameters.
Typically this means that metacells will be constructed separately for each biological
replicate, cell type or cell state, disease condition, etc. The metacell representation is
considerably less sparse than the original single-cell dataset, which is preferable for
co-expression network analysis or othter analyses that rely on correlations.
}
