% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ModuleConnectivity}
\alias{ModuleConnectivity}
\title{ModuleConnectivity}
\usage{
ModuleConnectivity(
  seurat_obj,
  harmonized = TRUE,
  assay = NULL,
  slot = "data",
  group.by = NULL,
  group_name = NULL,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Computes intramodular connectivity (kME) based on module eigengenes.
}
\examples{
ModuleConnectivity(pbmc)
}
\keyword{scRNA-seq}
