% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ModuleEigengenes}
\alias{ModuleEigengenes}
\title{ModuleEigengenes}
\usage{
ModuleEigengenes(
  seurat_obj,
  group.by.vars = NULL,
  modules = NULL,
  vars.to.regress = NULL,
  scale.model.use = "linear",
  verbose = TRUE,
  assay = NULL,
  pc_dim = 1,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{group.by.vars}{groups to harmonize by}

\item{modules}{table containing module / gene assignments, as in GetModules(seurat_obj).}

\item{vars.to.regress}{character vector of variables in seurat_obj@meta.data to regress when running ScaleData}

\item{scale.model.use}{model to scale data when running ScaleData choices are "linear", "poisson", or "negbinom"}

\item{verbose}{logical indicating whether to print messages}

\item{assay}{Assay in seurat_obj to compute module eigengenes. Default is DefaultAssay(seurat_obj)}

\item{pc_dim}{Which PC to use as the module eigengene? Default to 1.}

\item{wgcna_name}{name of the WGCNA experiment}
}
\description{
Computes module eigengenes for all WGCNA co-expression modules
}
\examples{
 ModuleEigengenes(pbmc)
}
\keyword{scRNA-seq}
