% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleEigengenes.R
\name{ModuleExprScore}
\alias{ModuleExprScore}
\title{ModuleExprScore}
\usage{
ModuleExprScore(
  seurat_obj,
  n_genes = 25,
  method = "Seurat",
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{n_genes}{the number of genes to use for each module, ranked by kME. Setting n_genes = 'all' uses all of the genes in a module}

\item{method}{selected method for module scoring, valid choices are "Seurat" or "UCell"}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\value{
seurat_obj with module scores computed for the selected wgcna experiment
}
\description{
Computes a module score for each co-expression module using Seurat AddModuleScore or UCell.
}
\details{
ModuleExprScore provides an alternative function to ModuleEigengenes for summarizing
the expression level of each module. The user can choose between Seurat AddModuleScore
or UCell using the method parameter.
}
