% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ModuleNetworkPlot}
\alias{ModuleNetworkPlot}
\title{ModuleNetworkPlot}
\usage{
ModuleNetworkPlot(
  seurat_obj,
  mods = "all",
  outdir = "ModuleNetworks",
  plot_size = c(6, 6),
  wgcna_name = NULL,
  label_center = FALSE,
  edge.alpha = 0.25,
  vertex.label.cex = 1,
  vertex.size = 6,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{mods}{Names of the modules to plot. If mods = "all", all modules are plotted.}

\item{outdir}{The directory where the plots will be stored.}

\item{plot_size}{A vector containing the width and height of the network plots.}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Visualizes the top hub genes for selected modules as a circular network plot
}
\examples{
ModuleNetworkPlot
}
\keyword{scRNA-seq}
