% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulePreservation.R
\name{ModulePreservationNetRep}
\alias{ModulePreservationNetRep}
\title{ModulePreservationNetRep}
\usage{
ModulePreservationNetRep(
  seurat_query,
  seurat_ref,
  name,
  n_permutations = 10000,
  n_threads = 8,
  gene_mapping = NULL,
  genome_ref_col = NULL,
  genome_query_col = NULL,
  TOM_use = NULL,
  wgcna_name = NULL,
  wgcna_name_ref = NULL,
  ...
)
}
\arguments{
\item{seurat_query}{A Seurat object}

\item{seurat_ref}{A Seurat object serving as the reference for the module preservation analysis}

\item{name}{The name to give the module preservation analysis.}

\item{n_permutations}{Number of permutations for the module preservation test.}

\item{n_threads}{Number of parallel threads to use for the module preservation test}

\item{gene_mapping}{a dataframe containing gene name mappings between the query and the reference dataset. One column should have the gene name in the query dataset, and anotehr column should have the corresponding gene name in the reference dataset.}

\item{genome_ref_col}{the column name containing the gene names for the reference dataset}

\item{genome_query_col}{the column name containing the gene names for the query dataset}

\item{TOM_use}{The name of the hdWGCNA experiment containing the TOM that will be used for plotting}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}

\item{wgcna_name_ref}{The name of the hdWGCNA experiment in the seurat_ref@misc slot}
}
\value{
A Seurat Object containing the module preservation statistics from NetRep.
}
\description{
Computes module preservation statistics in a query dataset for a given reference dataset
}
\details{
ModulePreservationNetRep performs a statistical test to assess the preservation of co-expression modules identified
in one dataset in an independent dataset using the NetRep implementation rather than the WGCNA implementation.
This method is originally described by Langfelder et al. in the 2011 paper
"Is My Network Module Preserved and Reproducible?". The NetRep implementation of the module preservation test
is described by Ritchie et al. in the 2016 paper "A Scalable Permutation Approach Reveals Replication and Preservation Patterns of Networks Modules in Large Datasets"
This method can be used to assess biological differences in networks, as well as technical differences / reproducibility across different
batches. Note that the outputs of ModulePreservation and ModulePreservationNetRep are not identical. This function requires separate
installation of the NetRep package, which is not directly included as a dependency in hdWGCNA.
}
