% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{ModuleTraitCorrelation}
\alias{ModuleTraitCorrelation}
\title{ModuleTraitCorrelation'}
\usage{
ModuleTraitCorrelation(
  seurat_obj,
  traits,
  group.by = NULL,
  features = "hMEs",
  cor_method = "pearson",
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A list of column names in the Seurat object's metadata that you wish to correlate with each module.
Traits must be a categorical variable (not a character vector), or a numeric variable.}

\item{features}{Which features to use to summarize each modules? Valid choices are hMEs, MEs, or scores}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}

\item{cor_meth}{Which method to use for correlation? Valid choices are pearson, spearman, kendall.}
}
\description{
Correlates categorical and numeric variables with Module Eigengenes or hub-gene scores.
}
\examples{
ModuleTraitCorrelation
}
\keyword{scRNA-seq}
