% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ModuleUMAPPlot}
\alias{ModuleUMAPPlot}
\title{ModuleUMAPPlot}
\usage{
ModuleUMAPPlot(
  seurat_obj,
  sample_edges = TRUE,
  edge_prop = 0.2,
  label_hubs = 5,
  edge.alpha = 0.25,
  vertex.label.cex = 0.5,
  label_genes = NULL,
  return_graph = FALSE,
  keep_grey_edges = TRUE,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{sample_edges}{logical determining whether we downsample edges for plotting (TRUE), or take the strongst edges.}

\item{edge_prop}{proportion of edges to plot. If sample_edges=FALSE, the strongest edges are selected.}

\item{label_hubs}{the number of hub genes to label in each module}

\item{edge.alpha}{scaling factor for edge opacity}

\item{vertex.label.cex}{font size for labeled genes}

\item{return_graph}{logical determining whether to plot thr graph (FALSE) or return the igraph object (TRUE)}

\item{keep_grey_edges}{logical determining whether to show edges between genes in different modules (grey edges)}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Makes a igraph network plot using the module UMAP
}
\examples{
ModuleUMAPPlot
}
\keyword{scRNA-seq}
