% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{OverlapBarPlot}
\alias{OverlapBarPlot}
\title{OverlapBarPlot}
\usage{
OverlapBarPlot(
  overlap_df,
  plot_var = "odds_ratio",
  logscale = FALSE,
  neglog = FALSE,
  label_size = 2,
  ...
)
}
\arguments{
\item{overlap_df}{the Module/DEG overlap table from OverlapModulesDEGs}

\item{plot_var}{the name of the overlap statistic to plot}

\item{logscale}{logical controlling whether to plot the result on a log scale, useful for odds ratio}

\item{neglog}{logical controlling wehether to plot the result as a negative log, useful for p-value / FDR}

\item{label_size}{the size of the module labels in the bar plot}
}
\description{
Plots the results from OverlapModulesDEGs as a bar plot
}
\examples{
OverlapBarPlot
}
\keyword{scRNA-seq}
