% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{OverlapDotPlot}
\alias{OverlapDotPlot}
\title{OverlapDotPlot}
\usage{
OverlapDotPlot(
  overlap_df,
  plot_var = "odds_ratio",
  logscale = TRUE,
  neglog = FALSE,
  plot_significance = TRUE,
  ...
)
}
\arguments{
\item{overlap_df}{the Module/DEG overlap table from OverlapModulesDEGs}

\item{plot_var}{the name of the overlap statistic to plot}

\item{logscale}{logical controlling whether to plot the result on a log scale, useful for odds ratio}

\item{neglog}{logical controlling wehether to plot the result as a negative log, useful for p-value / FDR}

\item{plot_significance}{logical controlling whether to plot the significance levels on top of the dots}
}
\description{
Makes barplots from Enrichr data
}
\examples{
OverlapDotPlot
}
\keyword{scRNA-seq}
