% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{OverlapModulesDEGs}
\alias{OverlapModulesDEGs}
\title{OverlapModulesDEGs}
\usage{
OverlapModulesDEGs(
  seurat_obj,
  deg_df,
  fc_cutoff = 0.5,
  group_col = "cluster",
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{deg_df}{DEG table formatted like the output from Seurat's FindMarkers}

\item{fc_cutoff}{log fold change cutoff for DEGs to be included in the overlap test}

\item{group_col}{the name of the column in deg_df containing the cell grouping information}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
Performs Fisher's Exact Test for overlap between DEGs and hdWGCNA modules.
}
\keyword{scRNA-seq}
