% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{OverlapModulesMotifs}
\alias{OverlapModulesMotifs}
\title{Overlap modules with TF target genes}
\usage{
OverlapModulesMotifs(seurat_obj, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Overlap modules with TF target genes
}
\examples{
OverlapModulesDEGs
}
\keyword{scRNA-seq}
