% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotDMEsVolcano}
\alias{PlotDMEsVolcano}
\title{PlotDMEsVolcano}
\usage{
PlotDMEsVolcano(
  seurat_obj,
  DMEs,
  plot_labels = TRUE,
  mod_point_size = 4,
  label_size = 4,
  show_cutoff = TRUE,
  wgcna_name = NULL
)
}
\arguments{
\item{DMEs}{dataframe output from FindDMEs or FindAllDMEs}

\item{plot_labels}{logical determining whether to plot the module labels}

\item{mod_point_size}{the size of the points in each plot}

\item{label_size}{the size of the module labels}

\item{show_cutoff}{logical determining whether to plot the significance cutoff. Should set this to FALSE if using facet_wrap.}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\value{
A ggplot object
}
\description{
Plotting function for the results of FindDMEs and FindAllDMEs
}
\examples{
PlotDMEsVolcano
}
\keyword{scRNA-seq}
