% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReassignModules.R
\name{ReassignModules}
\alias{ReassignModules}
\title{ReassignModules}
\usage{
ReassignModules(
  seurat_obj,
  harmonized = TRUE,
  features = NULL,
  new_modules = NULL,
  ignore = FALSE,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{harmonized}{logical indicating whether or not to use harmonized MEs}

\item{features}{character vector containing features for manual module reassignment,}

\item{new_modules}{character vector containing modules to reassign the genes}

\item{ignore}{logical indicating whether or not to ignore error message about reassigning non-grey features}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\value{
seurat_obj with the an updated modules table for the selected wgcna experiment
}
\description{
Reassigns features to different modules
}
\details{
ReassignModules reassigns features with negative kMEs in their assigned module to the
module that had the highest kME for that feature. Alternatively, this function
can manually assign features to different modules, which can be helpful if
certain genes of interest are assigned to the grey module. We generally do not
advise reassigning features to new modules if they are in non-grey modules,
but the user can do so at their own risk by setting ignore=TRUE.
}
