% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ROCCurves}
\alias{ROCCurves}
\title{ROCCurves}
\usage{
ROCCurves(seurat_obj, roc_df = NULL, conf_df = NULL, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{dbs}{List of EnrichR databases}

\item{max_genes}{Max number of genes to include per module, ranked by kME.}
}
\description{
Makes barplots from Enrichr data
}
\examples{
ROCCurves
}
\keyword{scRNA-seq}
