% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{RunModuleUMAP}
\alias{RunModuleUMAP}
\title{RunModuleUMAP}
\usage{
RunModuleUMAP(
  seurat_obj,
  features = "TOM",
  n_hubs = 10,
  exclude_grey = TRUE,
  wgcna_name = NULL,
  n_neighbors = 25,
  metric = "cosine",
  spread = 1,
  min_dist = 0.4,
  supervised = FALSE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{n_hubs}{number of hub genes to use in the UMAP computation}

\item{exclude_grey}{logical indicating whether to include grey module}

\item{wgcna_name}{The name of the scWGCNA experiment in the seurat_obj@misc slot}

\item{...}{Additional parameters supplied to uwot::umap}
}
\description{
Run UMAP on co-expression matrix using hub genes as features.
}
\keyword{scRNA-seq}
