% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModuleUMAP.R
\name{RunModuleUMAP}
\alias{RunModuleUMAP}
\title{RunModuleUMAP}
\usage{
RunModuleUMAP(
  seurat_obj,
  n_hubs = 10,
  exclude_grey = TRUE,
  genes_use = NULL,
  wgcna_name = NULL,
  n_neighbors = 25,
  metric = "cosine",
  spread = 1,
  min_dist = 0.4,
  supervised = FALSE,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{n_hubs}{number of hub genes to use in the UMAP computation}

\item{exclude_grey}{logical indicating whether to include grey module}

\item{genes_use}{character vector of genes to use for the UMAP, must already be present in GetModules(seurat_obj)}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}

\item{...}{Additional parameters supplied to uwot::umap}
}
\description{
Run UMAP on co-expression matrix using hub genes as features.
}
