% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacells.R
\name{RunPCAMetacells}
\alias{RunPCAMetacells}
\title{RunPCAMetacells}
\usage{
RunPCAMetacells(seurat_obj, wgcna_name = NULL, ...)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
Wrapper function to run Seurat's RunPCA function on the metacell object.
}
