% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetDatExpr}
\alias{SetDatExpr}
\title{SetDatExpr}
\usage{
SetDatExpr(
  seurat_obj,
  use_metacells = TRUE,
  wgcna_name = NULL,
  group.by = NULL,
  group_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}
}
\description{
This function sets up the expression matrix from the metacell object.
}
\examples{
SetDatExpr(pbmc)
}
\keyword{scRNA-seq}
