% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetModules}
\alias{SetModules}
\title{SetModules}
\usage{
SetModules(seurat_obj, modules, wgcna_name = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{modules}{dataframe containing gene module assignments}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\description{
SetModules
}
\keyword{scRNA-seq}
