% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_and_setters.R
\name{SetMultiExpr}
\alias{SetMultiExpr}
\title{SetMultiExpr}
\usage{
SetMultiExpr(
  seurat_obj,
  group_name,
  use_metacells = TRUE,
  group.by = NULL,
  multi.group.by = NULL,
  multi_groups = NULL,
  wgcna_name = NULL,
  assay = NULL,
  slot = "data",
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{group_name}{A string containing a group present in the provided group.by column or in the Seurat Idents.}

\item{use_metacells}{A logical determining if we use the metacells (TRUE) or the full expression matrix (FALSE)}

\item{group.by}{A string containing the name of a column in the Seurat object with cell groups (clusters, cell types, etc). If NULL (default), hdWGCNA uses the Seurat Idents as the group.}

\item{multi.group.by}{A string containing the name of a column in the Seurat object with groups for consensus WGCNA (dataset, sample, condition, etc)}

\item{multi_groups}{A character vecrtor containing the names of groups to select}

\item{wgcna_name}{A string containing the name of the WGCNA slot in seurat_obj@misc. Default = NULL which retrieves the currently active WGCNA data}
}
\description{
This function sets up the expression matrix input for consensus WGCNA based on
the metacell expression matrix, or from the full expression matrix.
}
\examples{
SetMultiExpr
}
\keyword{scRNA-seq}
