% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupForWGCNA.R
\name{SetupForWGCNA}
\alias{SetupForWGCNA}
\title{SetupForWGCNA}
\usage{
SetupForWGCNA(
  seurat_obj,
  wgcna_name,
  features = NULL,
  metacell_location = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{wgcna_name}{name of the WGCNA experiment}

\item{features}{list of features to use for WGCNA}

\item{metacell_location}{name of the WGCNA experiment to copy the metacell object from}

\item{...}{additional parameters to pass to SelectNetworkGenes}
}
\description{
Create a slot in a Seurat object to store hdWGCNA data
}
\details{
SetupForWGCNA creates a new slot in the Seurat object (seurat_obj) to store an hdWGCNA experiment
with a given name (wgcna_name). This function calls on SelectNetworkGenes to specify which features
will be used for network analysis. If there is another hdWGCNA experiment already in the seurat_obj,
the same metacell/metaspot object can be used by specifying the name of the hdWGCNA experiment
to the metacell_location parameter.
}
