% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_functions.R
\name{TestSoftPowers}
\alias{TestSoftPowers}
\title{TestSoftPowers}
\usage{
TestSoftPowers(
  seurat_obj,
  use_metacells = TRUE,
  group.by = NULL,
  group_name = NULL,
  powers = c(seq(1, 10, by = 1), seq(12, 30, by = 2)),
  make_plot = TRUE,
  outfile = "softpower.pdf",
  figsize = c(7, 7)
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{outfile}{filepath for output pdf to be generated. Default is}

\item{figsize}{numeric determining the height and width of the output figure. Default is c(7,7)}
}
\description{
This function gets the expression matrix from the metacell object.
}
\examples{
TestSoftPowers(pbmc)
}
\keyword{scRNA-seq}
