% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftPowers.R
\name{TestSoftPowers}
\alias{TestSoftPowers}
\title{TestSoftPowers}
\usage{
TestSoftPowers(
  seurat_obj,
  powers = c(seq(1, 10, by = 1), seq(12, 30, by = 2)),
  networkType = "signed",
  corFnc = "bicor",
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{powers}{numeric vector specifying soft powers to test}

\item{networkType}{The type of network to use for network analysis. Options are "signed" (default), "unsigned", or "signed hybrid". This should be consistent with the network chosen for ConstructNetwork}

\item{corFnc}{Correlation function for the gene-gene correlation adjacency matrix.}

\item{wgcna_name}{name of the WGCNA experiment}

\item{...}{additional parameters passed to WGCNA::pickSoftThreshold}
}
\description{
Compute the scale-free topology model fit for different soft power thresholds
}
\details{
TestSoftPowers aims to find the optimal value of the "soft-power threshold" for co-expression network analysis.
For varying values of this soft_power parameter, this function assesses the scale-free topology model fit of the resulting network
and the function also calculates other network statistics such as the connectivity. The user may also vary the type of network used and the correlation function.
For downstream functions, the lowest soft_power that reaches a fit of 0.8 is selected for network construction, but the user may select their own values.
}
