% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftPowers.R
\name{TestSoftPowersConsensus}
\alias{TestSoftPowersConsensus}
\title{TestSoftPowersConsensus}
\usage{
TestSoftPowersConsensus(
  seurat_obj,
  powers = c(seq(1, 10, by = 1), seq(12, 30, by = 2)),
  use_metacells = TRUE,
  networkType = "signed",
  corFnc = "bicor",
  setDatExpr = FALSE,
  group.by = NULL,
  group_name = NULL,
  multi.group.by = NULL,
  multi_groups = NULL,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{powers}{numeric vector specifying soft powers to test}

\item{use_metacells}{logical flag for whether to use the metacell expression matrix}

\item{networkType}{The type of network to use for network analysis. Options are "signed" (default), "unsigned", or "signed hybrid". This should be consistent with the network chosen for ConstructNetwork}

\item{corFnc}{Correlation function for the gene-gene correlation adjacency matrix.}

\item{setDatExpr}{logical flag indicating whether to run setDatExpr.}

\item{group.by}{A string containing the name of a column in the Seurat object with cell groups (clusters, cell types, etc). If NULL (default), hdWGCNA uses the Seurat Idents as the group.}

\item{group_name}{A string containing a group present in the provided group.by column or in the Seurat Idents. A character vector can be provided to select multiple groups at a time.}

\item{multi.group.by}{A string containing the name of a column in the Seurat object with groups for consensus WGCNA (dataset, sample, condition, etc)}

\item{multi_groups}{A character vecrtor containing the names of groups to select}

\item{...}{additional parameters passed to SetDatExpr}
}
\description{
Compute the scale-free topology model fit for different soft power thresholds separately for each input dataset
}
\examples{
# TestSoftPowers(pbmc)
}
\keyword{scRNA-seq}
